update to str_len (automated command)

+++ main.h
@@ -26,7 +26,7 @@ int string_printer(va_list string);
 unsigned long int _power(int base, int exponent);
 int number_len(long int n);
 int _putchar(char c);
-unsigned int _strlen(const char *s);
+unsigned long int _strlen(const char *s);
 int print_number(va_list number);
 void num_print_helper(int n, int *bytes);
 int _printf(const char *format, ...);
